+++ CHANGELOG.md
@@ -35,6 +35,7 @@ The following changes have been implemented but not released yet:
 #### browser and node
 
 - Fix the `Session` error listener typing by adding `Error` to the `errorDescription` type so that it reflects the actual behavior.
+  Thanks to @NoelDeMartin for fixing this issue. 
 
 ## [2.3.0](https://github.com/inrupt/solid-client-authn-js/releases/tag/v2.3.0) - 2024-11-14